# docs/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anychat_server_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_gateway_handler.UpdateProfileRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      birthday:
        example: "1990-01-01T00:00:00Z"
        type: string
      gender:
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      nickname:
        example: 张三
        type: string
      region:
        example: 北京
        type: string
      signature:
        example: 这是我的个性签名
        type: string
    type: object
...
  internal_gateway_handler.UserProfile:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      birthday:
        example: "1990-01-01T00:00:00Z"
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      gender:
        example: 1
        type: integer
      nickname:
        example: 张三
        type: string
      phone:
        example: "13800138000"
        type: string
      qrcode_url:
        example: https://example.com/qrcode.png
        type: string
      region:
        example: 北京
        type: string
      signature:
        example: 这是我的个性签名
        type: string
      user_id:
        example: user-123
        type: string
    type: object
  internal_gateway_handler.UserSearchItem:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      nickname:
        example: 张三
        type: string
      signature:
        example: 这是我的个性签名
        type: string
      user_id:
        example: user-123
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anychat.example.com
    name: AnyChat API Support
    url: https://github.com/yzhgit/anychat-server
  description: |-
    AnyChat instant messaging system gateway API service, providing HTTP interfaces for user authentication, user management, and other functions.
    All endpoints requiring authentication must include Authorization: Bearer <token> in the Header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AnyChat Gateway API
  version: "1.0"
paths:
...
  /users/{userId}:
    get:
      consumes:
      - application/json
      description: Get public info of specified user
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: get success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_gateway_handler.UserSearchItem'
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get user info
      tags:
      - user
  /users/me:
    get:
      consumes:
      - application/json
      description: Get detailed profile of currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: get success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_gateway_handler.UserProfile'
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get personal profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update personal profile of currently logged in user
      parameters:
      - description: profile info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: update success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_gateway_handler.UserProfile'
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: update personal profile
      tags:
      - user
  /users/me/email/bind:
    post:
      consumes:
      - application/json
      description: Bind email for currently logged in user
      parameters:
      - description: bind email info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.BindEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: bind success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "409":
          description: email already in use
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: bind email
      tags:
      - user
  /users/me/email/change:
    post:
      consumes:
      - application/json
      description: Change bound email for currently logged in user
      parameters:
      - description: change email info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.ChangeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: change success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "409":
          description: email already in use
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: change email
      tags:
      - user
  /users/me/phone/bind:
    post:
      consumes:
      - application/json
      description: Bind phone number for currently logged in user
      parameters:
      - description: bind phone info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.BindPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: bind success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "409":
          description: phone number already in use
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: bind phone
      tags:
      - user
  /users/me/phone/change:
    post:
      consumes:
      - application/json
      description: Change bound phone number for currently logged in user
      parameters:
      - description: change phone info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.ChangePhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: change success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "409":
          description: phone number already in use
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: change phone
      tags:
      - user
  /users/me/push-token:
    post:
      consumes:
      - application/json
      description: Update device push notification token
      parameters:
      - description: push token info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.UpdatePushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: update success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: update push token
      tags:
      - user
...
  /users/search:
    get:
      consumes:
      - application/json
      description: Search users by keyword (supports nickname, phone, user ID)
      parameters:
      - description: search keyword
        in: query
        name: keyword
        required: true
        type: string
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: search success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_gateway_handler.UserSearchResult'
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: search users
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"